fmpz_mod_poly get
